Replaced critical uses of double with BigDecimal

+++ src/main/java/info/blockchain/api/exchangerates/Currency.java
@@ -1,63 +1,57 @@
 package info.blockchain.api.exchangerates;
 
+import java.math.BigDecimal;
+
 /**
- *  Used in response to the `getTicker` method in the `ExchangeRates` class.
- *
+ * Used in response to the `getTicker` method in the `ExchangeRates` class.
  */
-public class Currency
-{
-	private double buy;
-	private double sell;
-	private double last;
-	private double price15m;
+public class Currency {
+	private BigDecimal buy;
+	private BigDecimal sell;
+	private BigDecimal last;
+	private BigDecimal price15m;
 	private String symbol;
-	
-	public Currency(double buy, double sell, double last, double price15m, String symbol)
-	{
-		this.buy = buy;
-		this.sell = sell;
-		this.last = last;
-		this.price15m = price15m;
+
+	public Currency (double buy, double sell, double last, double price15m, String symbol) {
+		this.buy = BigDecimal.valueOf(buy);
+		this.sell = BigDecimal.valueOf(sell);
+		this.last = BigDecimal.valueOf(last);
+		this.price15m = BigDecimal.valueOf(price15m);
 		this.symbol = symbol;
 	}
 
 	/**
 	 * @return Current buy price
 	 */
-	public double getBuy()
-	{
+	public BigDecimal getBuy () {
 		return buy;
 	}
 
 	/**
 	 * @return Current sell price
 	 */
-	public double getSell()
-	{
+	public BigDecimal getSell () {
 		return sell;
 	}
 
 	/**
 	 * @return Most recent market price
 	 */
-	public double getLast()
-	{
+	public BigDecimal getLast () {
 		return last;
 	}
 
 	/**
 	 * @return 15 minutes delayed market price
 	 */
-	public double getPrice15m()
-	{
+	public BigDecimal getPrice15m () {
 		return price15m;
 	}
 
 	/**
 	 * @return Currency symbol
 	 */
-	public String getSymbol()
-	{
+	public String getSymbol () {
 		return symbol;
 	}
 }
